requirements.txt
torch>=1.9.0,<2.0.0
pytorch-forecasting>=0.10.0,<0.11.0
pytorch-lightning>=1.5.0,<1.10.0
psycopg2-binary
pandas>=1.3.0
numpy>=1.21.0
scikit-learn
plotly
fastapi
uvicorn
python-dotenv
sqlalchemy
jupyter
matplotlib
seaborn